+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Enhancements 改进
 
 1. Show fetching status at footer play bar. 在播放底栏显示加载状态。
+2. Use HarmonyOS-Sans as default font on Windows. 在 Windows 使用 HarmonyOS-Sans 作为默认字体。
 
 # 1.1.0 (2026-01-26)